Refactor async stability tests
