+++ test/stability-tests/src/it/java/software/amazon/awssdk/stability/tests/s3/S3BaseStabilityTest.java
@@ -15,11 +15,18 @@
 
 package software.amazon.awssdk.stability.tests.s3;
 
+import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.CompletableFuture;
+import java.util.function.IntFunction;
+
+import org.apache.commons.lang3.RandomStringUtils;
+import software.amazon.awssdk.core.async.AsyncRequestBody;
+import software.amazon.awssdk.core.async.AsyncResponseTransformer;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.http.apache.ApacheHttpClient;
 import software.amazon.awssdk.services.s3.S3AsyncClient;
@@ -27,6 +34,7 @@ import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.DeleteBucketRequest;
 import software.amazon.awssdk.services.s3.model.NoSuchBucketException;
 import software.amazon.awssdk.services.s3.model.NoSuchKeyException;
+import software.amazon.awssdk.stability.tests.utils.StabilityTestRunner;
 import software.amazon.awssdk.testutils.RandomTempFile;
 import software.amazon.awssdk.testutils.service.AwsTestBase;
 import software.amazon.awssdk.utils.Logger;
@@ -50,6 +58,85 @@ public abstract class S3BaseStabilityTest extends AwsTestBase {
 
     protected String computeKeyName(int i) {
         return "key_" + i;
+    }
+
+    protected abstract S3AsyncClient getTestClient();
+
+    protected abstract String getTestBucketName();
+
+    protected void doGetBucketAcl_lowTpsLongInterval() {
+        IntFunction<CompletableFuture<?>> future = i -> getTestClient().getBucketAcl(b -> b.bucket(getTestBucketName()));
+        String className = this.getClass().getSimpleName();
+        StabilityTestRunner.newRunner()
+                .testName(className + ".getBucketAcl_lowTpsLongInterval")
+                .futureFactory(future)
+                .requestCountPerRun(10)
+                .totalRuns(3)
+                .delaysBetweenEachRun(Duration.ofSeconds(6))
+                .run();
+    }
+
+
+    protected void downloadLargeObjectToFile() {
+        File randomTempFile = RandomTempFile.randomUncreatedFile();
+        StabilityTestRunner.newRunner()
+                .testName("S3AsyncStabilityTest.downloadLargeObjectToFile")
+                .futures(getTestClient().getObject(b -> b.bucket(getTestBucketName()).key(LARGE_KEY_NAME),
+                        AsyncResponseTransformer.toFile(randomTempFile)))
+                .run();
+        randomTempFile.delete();
+    }
+
+    protected void uploadLargeObjectFromFile() {
+        RandomTempFile file = null;
+        try {
+            file = new RandomTempFile((long) 2e+9);
+            StabilityTestRunner.newRunner()
+                    .testName("S3AsyncStabilityTest.uploadLargeObjectFromFile")
+                    .futures(getTestClient().putObject(b -> b.bucket(getTestBucketName()).key(LARGE_KEY_NAME),
+                            AsyncRequestBody.fromFile(file)))
+                    .run();
+        } catch (IOException e) {
+            throw new RuntimeException("fail to create test file", e);
+        } finally {
+            if (file != null) {
+                file.delete();
+            }
+        }
+    }
+
+    protected void putObject() {
+        byte[] bytes = RandomStringUtils.randomAlphanumeric(10_000).getBytes();
+
+        IntFunction<CompletableFuture<?>> future = i -> {
+            String keyName = computeKeyName(i);
+            return getTestClient().putObject(b -> b.bucket(getTestBucketName()).key(keyName),
+                    AsyncRequestBody.fromBytes(bytes));
+        };
+
+        StabilityTestRunner.newRunner()
+                .testName("S3AsyncStabilityTest.putObject")
+                .futureFactory(future)
+                .requestCountPerRun(CONCURRENCY)
+                .totalRuns(TOTAL_RUNS)
+                .delaysBetweenEachRun(Duration.ofMillis(100))
+                .run();
+    }
+
+    protected void getObject() {
+        IntFunction<CompletableFuture<?>> future = i -> {
+            String keyName = computeKeyName(i);
+            Path path = RandomTempFile.randomUncreatedFile().toPath();
+            return getTestClient().getObject(b -> b.bucket(getTestBucketName()).key(keyName), AsyncResponseTransformer.toFile(path));
+        };
+
+        StabilityTestRunner.newRunner()
+                .testName("S3AsyncStabilityTest.getObject")
+                .futureFactory(future)
+                .requestCountPerRun(CONCURRENCY)
+                .totalRuns(TOTAL_RUNS)
+                .delaysBetweenEachRun(Duration.ofMillis(100))
+                .run();
     }
 
     protected static void deleteBucketAndAllContents(S3AsyncClient client, String bucketName) {
