+++ test/stability-tests/src/it/java/software/amazon/awssdk/stability/tests/s3/S3NettyAsyncStabilityTest.java
@@ -10,7 +10,7 @@ import software.amazon.awssdk.stability.tests.utils.RetryableTest;
 
 import java.time.Duration;
 
-public class S3NettyAsyncStabilityTest extends S3AsyncStabilityTest {
+public class S3NettyAsyncStabilityTest extends S3BaseStabilityTest {
 
     private static String bucketName = "s3nettyasyncstabilitytests" + System.currentTimeMillis();
